Update reorganized.wl
removes choosing omega and Ve. Instead, asks for an energy in GeV for the vacuum, and calculates the Ve from the data automatically.

Added in association for the radii.

Moved where some things are stored and returned.  particularly, ktarget and the radial self interaction potential are now computed earlier, in SCalcScale and stabilityMatrix respectively.

updated the evscale function. It now computes the eigenvalues and returns them for a data set and some inputers.

SCalcScale now computers the eigenvalue list, potential, matrix, and ktarget for a particular radius.

Added buildkGrid, which builds the log spaced k grid around k target for the given infile, inputs, and number of steps.

Added kAdapt, which runs over multiple radii and for the correct adaptive kgrid, returning a list of points in the formater {raidus (cm),k (erg), Largest Im[eigenvalue (erg), Hsi radial at r (erg)]}

Still requires comparsion test of results to working code, and some speed tests are needed (it *feels* slower than the working version).