ndensity function & box fit intialguess
Added a function which computes the mu matricies for nu, nubar, and nux neutrino flavors.  This is the only place where the "xflavor" option makes an actual change, as whether mux is zero or nonzero propogates through.  The option is still required on most functions.  Seperated out the intial guess for the box fit into ti's own function.